new speed calculations

+++ project/source/game/entity/ItemSlot.cpp
@@ -163,8 +163,8 @@ void GetItemString(string& str, const Item* item, Unit* unit, uint count)
 			else
 				armor_type = Format("%s (%s)", skill, txInvalidArmor);
 
-			int old_mob = unit->CalculateMobility();
-			int new_mob = unit->CalculateMobility(&armor);
+			int old_mob = (int)unit->CalculateMobility();
+			int new_mob = (int)unit->CalculateMobility(&armor);
 			if(old_mob == new_mob)
 				mob_str = Format("(%d)", new_mob);
 			else
